soc: qcom: ipa: move code into rmnet probe and remove functions
Initialization of the IPA RMnet driver begins when ipa3_post_init()
calls ipa3_wwan_init().  That function registers the driver, which
will lead to the probe function being called.  Shutdown of the
driver is similar, with ipa3_wwan_cleanup() unregistering the
driver, which will lead to its remove function being called.

There is a some initialization that occurs in ipa3_wwan_init() that
can instead be done in the probe routine.  Moving that code there
means we can assume rmnet_ipa3_ctx points to a completely zeroed
structure, and that the probe function initializes everything for
the driver.

Similarly, a few calls in ipa3_wwan_cleanup() can be moved to the
remove function, meaning all cleanup for the driver is done once
the remove function has executed.

Signed-off-by: Alex Elder <elder@linaro.org>