soc: qcom: ipa: rmnet_ipa3_ctx points to zeroed data
We know that the structure rmnet_ipa3_ctxt refers to will be
zero-filled initially (it's in the BSS section), and that it will be
zeroed if the driver is ever removed.  It is also zeroed if
ipa3_wwan_probe() doesn't complete successfully.

Knowing the context will be zero-filled when ipa3_wwan_probe() is
called, we can skip the zero-initialization of some fields.

Signed-off-by: Alex Elder <elder@linaro.org>