soc: qcom: ipa: use dev in apps_ipa_tx_complete_notify()
Two spots in apps_ipa_tx_complete_notify() use wwan_ptr->net to get
a device pointer to operate on.  The value of wwan_ptr is retrieved
from netdev_priv(dev).  We know that the value of dev is the right
netdev pointer (it's even checked at the top of that function).
And we wwan_ptr->net is assigned in ipa3_wwan_probe() to point to
the same netdev pointer.  So just use dev in place of wwan_ptr->net
in those two spots.

Signed-off-by: Alex Elder <elder@linaro.org>