soc: qcom: ipa: define smp2p information in main IPA node
There is no need to have the information about smp2p signals used by
IPA be associated with a distinct device.  Just move those properties
into the main IPA node.  Also move the hardware and ee value
definitions earlier in that node.

Rather than probing for the sub-device, just initialize the smp2p
information directly.  Rename the smp2p initialization function
ipa_smp2p_init() to make it obvious it is no longer used as a
device probe function.

Create a new function, ipa_smp2p_exit(), which reverses the effect
of ipa_smp2p_init().  All it needs to do is zero the smp2p field in
the IPA context; the interrupt is a managed device resource so will
be released automatically.

Call ipa_smp2p_init() early in the main IPA probe function, to avoid
having to re-initialize things in the event it returns -EPROBE_DEFER.
(It depends on the smp2p driver, which may not be ready when we are
probed.)

This leaves no more devices subsidiary to the main IPA device, so
remove the call to of_platform_populate() at the end of the main
device probe routine.

Signed-off-by: Alex Elder <elder@linaro.org>