soc: qcom: ipa: fix a bug in ipa3_freeze_clock_vote_and_notify_modem()
If either the modem or the AP crash, the IPA code on the AP tells
the modem whether it had the clock enabled (i.e., whether it has
cast a vote for IPA hardware) using a smp2p signal.  Two bits are
used: the low-order bit indicates whether the next bit is valid (0
means no, 1 means yes); and the next bit indicates whether the clock
is on (0 means no).

The first problem is that the position of the bit representing
whether the clock is enabled should be 1, but in "sdm845.dtsi"
it is specified as 0.

The second (more important) problem is that the code that calculates
the value to write is completely bogus.

Correct the value in the DT file, and fix the bug.

Signed-off-by: Alex Elder <elder@linaro.org>