soc: qcom: ipa: signal valid separate from enabled in smp2p
The AP communicates two significant bits when informing the modem
about the state of the IPA clock.  Update the valid bit separate
from the clock enabled bit in the smp2p channel.

Note:  This is a bit strange, because we're reusing the state
pointer to update a bit that differs from the one that was initially
fetched from the DTB.  The smp2p implementation permits this (and
is likely to be changed soon to have a more coherent API).

Signed-off-by: Alex Elder <elder@linaro.org>