soc: qcom: ipa: fix QMI code
In testing, we found that the QMI code couldn't synchronously wait
for the response to a request if the request was issued in the
new_server handler function.  So we now send the INIT_DRIVER request
only, and arrange for a separate handler function for the response
to that request.  There are a few other changes in here, but they're
all fixes, and make this code match what has been tested.

Signed-off-by: Alex Elder <elder@linaro.org>