soc: qcom: ipa: don't use workqueue for phase 2 initialization
Currently arrangements to start the second initialization phase
(ipa3_post_init()) are made by scheduling it to occur via workqueue.

The reason it was scheduled was to avoid having the user space
process hang if there's a problem accessing the IPA hardware.
Outside of that, there's really no need for this.  The call is made
within ipa3_write(), which itself will be running in user context.

Just call ipa3_post_init() directly, rather than scheduling it to be
run via workqueue.

Signed-off-by: Alex Elder <elder@linaro.org>