soc: qcom: ipa: make GSI responsible for channel memory
Move responsibility for allocating and freeing DMA memory used for
the channel ring out of the IPA code and into the GSI code.
So gsi_alloc_channel() will now allocate the DMA memory used for the
channel ring, and gsi_dealloc_channel() will free it.

This makes it no longer necessary to keep a redundant copy of the
memory information in ipa3_ep_context->gsi_chan_ring_mem.

Signed-off-by: Alex Elder <elder@linaro.org>