soc: qcom: ipa: hide gsi_ctx type
The only references to the gsi_ctx global is now inside "gsi.c", so
that symbol can be declared static there.  The only uses of the
gsi_ctx type outside "gsi.c" are as a pointer.  So the struct type
can be declared externally where it's needed to define a pointer
field, but defined only within "gsi.c".

Signed-off-by: Alex Elder <elder@linaro.org>