soc: qcom: ipa: use common return in ipa3_reset_gsi_channel()
Use the same "result" variable to hold the return value, and use a
common return path in ipa3_reset_gsi_channel() whether there is an
error or not.  The sole caller of this function doesn't care what
particular non-zero value is returned, so don't bother changing it
to be -EFAULT if gsi_reset_channel() fails.

Signed-off-by: Alex Elder <elder@linaro.org>