soc: qcom: ipa: check event_id earlier in ipa_gsi_irq_rx_notify_cb()
If an unexpected event id is supplied to ipa_gsi_irq_rx_notify_cb()
it will report an error and return early.

Currently, it does so after verifying that the received packet was
the next one expected, and after recording the number of bytes that
were received.

I assert that if the event received is not expected, we should not
be recording whatever event length got recorded with the event.
So we should move the event type check to happen right away.  That's
what this patch does.

Signed-off-by: Alex Elder <elder@linaro.org>