soc: qcom: ipa: cleanup after last patch
Nothing ever waits for the general endpoint command completion, so
it can be removed.  That makes gsi_handle_gp_int1() a no-op
function; remove that and report a message in the one place it was
previously called.  (It may be that without issuing general
immediate commands we'll never see the GP_INT1 interrupt and it
should be treated as an bug.  But I don't know this, so we'll treat
it as an error instead.)

In addition, nothing ever updates the halt_channel field of the
gsi_generic_ee_cmd_debug_stats structure any more, and that was the
only field contained in that structure.  So we can remove its
definition, and delete the gen_ee_cmd_dbg field of the GSI context.

Without the GSI_GEN_EE_CMD_HALT_CHANNEL ever being used, there is no
need for the gsi_generic_ee_cmd_opcode enumerated type, so get rid
of that.

Finally, the GSI_EE_n_GSI_EE_GENERIC_CMD_OFFS() macro is no
longer used, so its definition, as well as that of its field bitmask
and maximum N value can be removed.

Signed-off-by: Alex Elder <elder@linaro.org>