soc: qcom: ipa: kill gsi_evt_ctx->scratch
Nothing ever writes to the scratch field of the gsi_evt_ctx
structure.  It will always be zero (the value it takes on when
the context is initialized in gsi_alloc_evt_ring().

Make this explicit by having __gsi_write_evt_ring_scratch() simply
write zeroes to event ring scratch 0 and 1 registers.  We no longer
need the "val" argument, so remove that.  Rename the function to
indicate it is zeroing the scratch registers rather than writing a
particular value.

This makes the scratch field of the gsi_evt_ctx structure unneeded,
so remove it, as well as the definition of its structure type.
Doing that eliminates the last use of the gsi_mhi_evt_scratch and
gsi_xdci_evt_scratch structures, so they too can go away.

Signed-off-by: Alex Elder <elder@linaro.org>