soc: qcom: ipa: channel is allocated in gsi_write_channel_scratch()
There is only one caller of gsi_write_channel_scratch(), and it
follows immediately after a successful call to gsi_alloc_channel().
When gsi_alloc_channel() succeeds, the state of the channel will
always be GSI_CHAN_STATE_ALLOCATED.  So there's no need to checm
the state of the channel at the top of gsi_write_channel_scratch().

Signed-off-by: Alex Elder <elder@linaro.org>