soc: qcom: ipa: initialize scratch info in __gsi_write_channel_scratch()
The only place a channel context's scratch field is written is in
gsi_write_channel_scratch().  That happens immediately after the
channel is allocated.  As soon as the field is updated, a call to
__gsi_write_channel_scratch() writes the values from that field
into the scratch registers.

There are two other callers of __gsi_write_channel_scratch(), but
they don't change the values held in the channel's scratch field.
So the values written by __gsi_write_channel_scratch() can just as
well be computed in that function.

Signed-off-by: Alex Elder <elder@linaro.org>