soc: qcom: ipa: gsi_config_channel_mode() return can be ignored
The only time gsi_config_channel_mode() returns an error is if the
channel is already in the requested mode (polling or callback).

Only one caller of gsi_config_channel_mode() checks its return
value:  ipa3_rx_switch_to_intr_mode().  Just prior to the call, it
will have done an atomic exchange of the current polling state and
determined that it was polling.  There is no reason to believe the
channel's polling mode differs from the endpoint's polling state,
nor that the channel mode would change between the exchange
operation and the call to gsi_config_channel_mode().

So have ipa3_rx_switch_to_intr_mode() ignore the return value of
gsi_config_channel_mode().  This allows for the logic in the calling
function to get simplified a bit.

Signed-off-by: Alex Elder <elder@linaro.org>