soc: qcom: ipa: kill enum gsi_chan_mode
The only remaining user of the gsi_chan_mode enumerated type is
gsi_alloc_channel().  Rather than setting the channel context
poll_mode field to GSI_CHAN_MODE_CALLBACK, just set it explicitly
to 0.  The field name suggests that 0 means "not polling."

With that, we can remove the definition of gsi_chan_mode.

Signed-off-by: Alex Elder <elder@linaro.org>