soc: qcom: ipa: kill enum gsi_intr_type
The gsi_intr_type enumerated type defines the two possible types of
interrupts that an event reing can use, MSI or IRQ.  This can easily
be represented by Boolean values, without having to define a special
type for this purpose.  Variable names can convey the meaning as
easily as the specific values can, and the places where actual
values that reach hardware are computed are very localized.

Signed-off-by: Alex Elder <elder@linaro.org>