soc: qcom: ipa: kill enum gsi_chan_dir
The two possible values of the gsi_chan_dir enumerated type can
easily be represented by a Boolean.  The only place that type was
used was in recording the direction of a channel.  Replace the
"dir" field of the gsi_chan_props structure with a Boolean
"from_gsi", which indicates by its name the direction based on its
value.  The only place values that reach hardware are calculated
based on this is in gsi_program_chan_ctx(); update that to supply
the proper numeric value based on the value of "from_gsi".

This allow us to eliminate the definition of gsi_chan_dir.

Signed-off-by: Alex Elder <elder@linaro.org>