soc: qcom: ipa: disallow EOB in ipa_gsi_irq_rx_notify_cb()
If a packet trasferred is held in multiple buffers, the completion
of an individual buffer operation is signaled by EOB, while the
completion of the entire transfer is signaled by EOT.

We only ever send single-buffer transfers, and as a result we only
ever expect an EOT event in ipa_gsi_irq_tx_notify_cb().

Currently, in the RX path ipa_gsi_irq_rx_notify_cb() allows either
EOT or EOB events.  But we should only receive single-buffer packets
as well, so receiving EOB should be treated as unexpected.

Stop accepting EOB events in ipa_gsi_irq_rx_notify_cb(), and report
it if one is ever seen.

Signed-off-by: Alex Elder <elder@linaro.org>