soc: qcom: ipa: kill gsi_max_channel_get()
Check in gsi_alloc_channel() whether the channel id it is supplied
is in range.  Note that we do it outside gsi_validate_channel_props()
because gsi_set_channel_cfg() (the other caller of that function)
supplies the same channel id as was set when the channel was
initially allocated; so there's no need to check it again.

With this check in place, we no longer need to check for a bad
channel number in ipa_gsi_setup_channel().  And we therefore have no
more need for gsi_max_channel_get(), so get rid of that.

Signed-off-by: Alex Elder <elder@linaro.org>