soc: qcom: ipa: we are not in poll mode in chan_xfer_cb()
In chan_xfer_cb() there is a check to see if the channel's poll_mode
flag is set, and if so, a warning message is issued.

There is no need for this.  Here's why.
  - chan_xfer_cb() is only called by gsi_process_chan(), and then only
    when its callback argument is true
  - gsi_process_chan() is only called by gsi_process_evt_re(), which
    passes its callback argument to gsi_process_chan()
  - gsi_process_evt_re() is called in two places, and of those, only
    handle_event() passes true as the callback argument
  - handle_event() is called only by gsi_handle_ieob(), which is
    called only by gsi_isr()--the interrupt handler
Therefore, if handle_event is called for an event ring its interrupt
must have been enabled.

Continuing:
  - after initialization, an event's interrupt is only enabled or
    disabled by gsi_config_channel_mode()
  - other than initialization, a channel context's poll_mode field
    is only set by gsi_config_channel_mode()
  - the interrupt state and poll_mode are both updated under
    protection of a GSI context spinlock
  - when gsi_config_channel_mode() returns, either:
      - the channel context poll_mode will be polling and its event
        ring interrupt will be disabled; or
      - the channel context poll_mode will not be polling and its event
        ring interrupt will be enabled.
  - currently every event is associated with exactly one channel
Therefore if an event ring interrupt is enabled, its corresponding
channel is not in poll mode.

So to summarize, we can assume the channel mode is polling when
chan_xfer_cb() is called, and we can get rid of the check.

Signed-off-by: Alex Elder <elder@linaro.org>