soc: qcom: ipa: remaining symbol version fixes
Remove remaining versions embedded within symbol names.  The only
version we support is v3.5.1 (and v1.3 of GSI).  There remain a few
comments that talk about things supported (or not) starting at a
particular version of the hardware that remain.  When we add support
for new hardware versions that require different values, we will
update things accordingly.

Signed-off-by: Alex Elder <elder@linaro.org>