soc: qcom: ipa: fix resource group array indices
The indices on the resource group arrays were set up a long time
ago, to make the arrays a bit more readable.  Because of the way
the loops in ipa_set_resource_groups_min_max_limits() are done,
some of the entries were never touched for version 3.5.1 of IPA.
At the time, I used the version 3.0 index values, because that way
every entry in the array (including those never touched for 3.5.1)
would have an explicit index.

Now we can clean this all up.  We only support v3.5 values, and so
we can use the 3.5 symbols to index the resource group arrays.
Delete definitions like IPA_v3_5_MHI_GROUP_PCIE (which redefines
IPA_v3_5_GROUP_LWA_DL), because they are never used.  We can also
remove the entries that will never be touched; since we only support
one hardware version currently there is no need for the arrays to be
any bigger.

Delete all the v3.0 and v4.0 symbols, which are (now) never used.

Signed-off-by: Alex Elder <elder@linaro.org>