soc: qcom: ipa: locks and barriers
The checkpatch.pl script wants a comment to be included with the
definition of every spinlock and mutex definition.  It wants
something similar for any explicit barriers.

In the interest of getting this committed as soon as possible, this
patch simply adds a bogus comment at each of these locations.
Before sending the code out for upstream review, all of these will
be replaced with comments that supply actual information.

Signed-off-by: Alex Elder <elder@linaro.org>