soc: qcom: ipa: bit-field grouping and alignment in gsi.c
Align the colon preceding the widths of C bit-fields defined in
"gsi.c" on a common tab stop, and put a space between the colon and
the width.  Group each set of bit-fields that together comprise bits
sufficient to fill their "base object," so it's more obvious where
the boundaries between the base objects are.

Signed-off-by: Alex Elder <elder@linaro.org>