soc: qcom: ipa: remove a barrier in gsi_alloc_evt_ring()
There are two barrier calls in gsi_alloc_evt_ring() in the error
path.  They're there to ensure that a cleared bit in the event bitmap
is visible to other processors before returning from the function.

Previously a mutex was used, but this commit replaced it with these
barrier calls:
    msm: gsi: use barriers rather than mutex for clear_bit()

The first barrier isn't really needed.  The bit, previously set in
the GSI context event, mask will already be visible to other
processors because the mutex that protected it will guarantee it.
Nothing other than that mask field needs to be made visible, so
there's no need for the barrier.

We could probably do with the second one safely as well, but having
it there ensures avoids a case where a racing processor finds no
free event to use (despite this one being available).

Add comments to explain some information about the barrier.

Signed-off-by: Alex Elder <elder@linaro.org>