soc: qcom: ipa: move client add/remove calls into ipa_cfg_ep()
There are a number of functions that ipa_cfg_ep() calls, each of
which acquires and releases a client reference while writing a
register.  Instead, move all of those into ipa_cfg_ep() itself,
surrounding each of the calls it makes.  This makes all of the extra
reference count busywork obvious; this is cleaned up in the next
patch.

Signed-off-by: Alex Elder <elder@linaro.org>