soc: qcom: ipa: kill ipa_ep_cfg->seq
The seq field of the ipa_ep_cfg structure is never set; its fields
are always zero.

The only function that uses this structure is ipa_cfg_ep_seq(),
which only uses the field's seq_type field if its set_dynamic field
is true.  We know it will be false, so we can simplify that function
by no longer supporting the ability to set the sequencer type
dynamically.

That makes the seq_cfg argument to ipa_cfg_ep_seq() unnecessary, so
we can remove that, and that makes seq field of the ipa_ep_cfg
structure unused, so we can remove it completely.  That leaves the
ipa_ep_cfg_seq type unused, so we can do away with that as well.

Signed-off-by: Alex Elder <elder@linaro.org>