soc: qcom: ipa: sequencer type is not INVALID in ipa_cfg_ep_seq()
Only ipa_cfg_ep() calls ipa_cfg_ep_seq(), and it is only called when
the client is a producer.  The client handle passed is the same as
the client handle it receives as argument.

The only caller of ipa_cfg_ep() is ipa_setup_sys_pipe(), and it
provides the value returned by ipa_get_ep_mapping() as the client
handle.  That handle is the mapped value for the client field of
the sys_in structure ipa_setup_sys_pipe() is provided.

There are four callers of ipa_setup_sys_pipe(), and two of them
provide producers as their sys_in.client values.  The producers
are IPA_CLIENT_APPS_CMD_PROD and IPA_CLIENT_APPS_WAN_PROD.

The type used in ipa_cfg_ep_seq() comes from the sequencer_type
field of the client's endpoint configuration, found in the array
ipa_ep_configuration[][].  Looking at that array, the two clients
in question have these values for sequencer_type:
  APPS_CMD_PROD: IPA_DPS_HPS_SEQ_TYPE_DMA_ONLY
  APPS_WAN_PROD: IPA_DPS_HPS_SEQ_TYPE_2ND_PKT_PROCESS_PASS_NO_DEC_UCP

Neither of these is IPA_DPS_HPS_SEQ_TYPE_INVALID, so there's no need
to check for that condition in ipa_cfg_ep_seq().

Signed-off-by: Alex Elder <elder@linaro.org>