soc: qcom: ipa: drop an assertion in ipa_cfg_ep_seq()
There is an assertion at the top of ipa_cfg_ep_seq(), ensuring that
the client's sequencer type is a DMA type if the endpoint mode was
DMA.

The previous patch explained that only APPS_CMD_PROD and APPS_WAN_PROD
are supplied to ipa_cfg_ep_seq() as the client handle.  The former
has its input parameters set up in setup_apps_cmd_prod_pipe(), where
it sets the mode to IPA_DMA.  The latter's input parameters are set
up in handle3_egress_format(), where the mode is set to IPA_BASIC.

So the APPS_CMD_PROD case is subject to the assertion.  But as the
previous patch indicated, that client has sequencer type
IPA_DPS_HPS_SEQ_TYPE_DMA_ONLY, which *is* a DMA type.  I.e.:
   ipa_dps_hps_seq_type_is_dma(type) is true

So we can get rid of this test and assertion.  This removes the only
use of ipa_dps_hps_seq_type_is_dma(), so we must remove that
function too.

Doing this assertion is reasonable to do, the mode *can* be set up
dynamically and could in principle be configured with a value
incompatible with the sequencer type.  But right now we know
the only place DMA mode is used, it has the right type.

Signed-off-by: Alex Elder <elder@linaro.org>