soc: qcom: ipa: kill gsi_irq_update()
The only caller of gsi_irq_update() is gsi_irq_control_event(), and
the way it is used isn't as clear as it should be.

Remove gsi_irq_update(), and implement gsi_irq_control_event() in a
way that makes it more obvious it's just turning on or off a bit
corresponding to an event.

Signed-off-by: Alex Elder <elder@linaro.org>