soc: qcom: ipa: simplify gsi_process_evt_re()
The index of an event ring entry local read pointer is computed by
ring_rp_local_index().  It does this by subtracting the base from
the rp_local pointer, and dividing by the element size.  Right after
gsi_process_evt_re() gets this information, it computes the address
of the event at that index in the ring.  This is done by multiplying
that index by the element size and adding that to the base address.

Rather than doing all this math, just get the event pointer by
converting the rp_local physical address to its virtual equivalent.

Signed-off-by: Alex Elder <elder@linaro.org>