soc: qcom: ipa: ipa_cfg_ep_ctrl() is used only on consumer pipes
The only caller of ipa_cfg_ep_ctrl() is suspend_pipe(), and the only
caller of suspend_pipe() passes only consumer pipes (LAN and WAN) as
the client.  So there's no need for ipa_cfg_ep_ctrl() to ensure the
pipe is a consumer before calling ipa_suspend_active_aggr_wa().
Add some brief comment to indicate that it is expected we're
operating with a consumer client.

Signed-off-by: Alex Elder <elder@linaro.org>