soc: qcom: ipa: wait to disable GSI event interrupt
In ipa_gsi_poll_after_suspend(), the interrupt for a channel is
disabled unconditionally, before determining whether the endpoint is
in polling state.

Instead, disable interrupts after the atomic operation sets the
endpoint into polling state.  That way curr_polling_state controls
whether interrupts are disabled via atomic operations.

Interrupts are enabled in ipa_rx_switch_to_intr_mode(), again under
protection of an atomic update of the curr_polling_state value.

The only time an event's interrupt state is changed without having
this protection is when a pipe is resumed, where interrupts are
enabled unconditionally if the endpoint was not in polling mode.
I assume this means nothing else will execute this path
concurrently, so doing this is safe.  (And even thought it's not
adequate protection, it *does* check curr_polling_state when
determining whether to enable the interrupt.)

Signed-off-by: Alex Elder <elder@linaro.org>