soc: qcom: ipa: always start NAPI poll via workqueue
In ipa_gsi_irq_rx_notify_cb(), non-NAPI channels continue their work
by scheduling their sys->work (ipa_wq_handle_rx()) to be run.  The
result is that ipa_handle_rx() is called for the channel, in
workqueue context.

For NAPI channels, we do a non-blocking attempt to get a client
reference, and if that's successful we call the endpoint's
client->notify function, signaling to start polling.  Otherwise the
work function is scheduled to be run (because this function is only
ever called in interrupt context).  The work function is again
ipa_wq_handle_rx(), and all it does is acquire the client reference
(blocking if needed) and then calls the client->notify function.

The client notify is set from the sys_in.notify value supplied to
ipa_setup_sys_pipe().  APPS_WAN_CONS is the only NAPI consumer pipe
we have, and its notify function is apps_ipa_packet_receive_notify().
When it gets a IPA_CLIENT_START_POLL notification, it simply calls
napi_schedule().

This patch skips over the non-blocking attempt to get a client
reference, instread just leaving it up to ipa_wq_handle_rx() to get
the reference and call the notify function.

This change could add a small bit of latency before the first NAPI
poll call if the IPA clock was already running.  But some future
changes are expected to eliminate that.

Signed-off-by: Alex Elder <elder@linaro.org>