soc: qcom: ipa: assume good packet in ipa_gsi_irq_rx_notify_cb()
The xfer_data argument to ipa_gsi_irq_rx_notify_cb() provides the
address of the TX packet wrapper that is next expected to arrive.
This is used for a sanity check.  If it is *not* the next one,
a warning is issued, but otherwise the completion event is ignored.

This patch deletes the sanity check.  If the pointer supplied is not
the next TX wrapper, we don't know *what* it is.  And if that is
wrong there's no reason to believe the next event will supply a
valid wrapper either.

Signed-off-by: Alex Elder <elder@linaro.org>