soc: qcom: ipa: kill ipa_ep_cfg->conn_track
The conn_track field of the ipa_ep_cfg structure is never used, so
get rid of it.  This was the only use of the ipa_ep_cfg_conn_track
structure type, so that can also be removed.  And that removes the
need for the ipa_ipv6ct_en_type enumerated type, so get rid of that
too.

Signed-off-by: Alex Elder <elder@linaro.org>