soc: qcom: ipa: kill ipa_ep_cfg->nat
The nat field of the ipa_ep_cfg structure is never used, so we can
get rid of it.  This field was the only place the ipa_ep_cfg_nat
structure was required, but it can't be removed without removing the
construct function for the IPA_ENDP_INIT_NAT_n register.  That too
can be removed because--like the nat field--nothing ever uses that
register.  So remove the construct function, and the entry in the
ipahal_reg_objs[][] array that references it, and the definition of
IPA_ENDP_INIT_NAT_n symbol in the ipahal_reg enumerated type.

Signed-off-by: Alex Elder <elder@linaro.org>