soc: qcom: ipa: let compiler initialize local structs
Rather than calling:
  memset(&foo, 0, sizeof(foo));
to zero local structure variables, initialize them when declared:
  struct foo foo = { };

Replace uses of "{ 0 }" to do this sort of initialization in some
spots with just "{ }".  The latter works for all structures, while
the former can't be used when structures have nested struct fields.

Doing this made it obvious that a variable in ipa_stop_gsi_channel()
was never used.

Signed-off-by: Alex Elder <elder@linaro.org>