soc: qcom: ipa: don't zero array in ipa_init_mem_info()
There is no need to zero the array whose address is passed to
ipa_init_mem_info(); it will already be zeroed (because it's
embedded in the IPA context, which is initially zero-filled).

If an error occurs in ipa_plat_drv_probe() after the array has been
initialized, zero it before returning, so to return to our initial
state.

Signed-off-by: Alex Elder <elder@linaro.org>