soc: qcom: ipa: remove unused memory partition info
Remove the names of all memory partition slots that are never used
in the IPA code, as well as the values used to initialize those
slots.

Some remaining slots are only used to set "canary" values in memory
between regions.  I was told that other code (e.g., the modem) might
depend on these, so they must remain.

Signed-off-by: Alex Elder <elder@linaro.org>