soc: qcom: ipa: kill ALIGN_CHECK()
The ALIGN_CHECK() macro encapsulates verifying at compile time that
the alignment of an offset value adheres to its requirements, while
assigning the value.  All but one of the alignment requirements are
8, and the other is 1K.  Having a symbolic constant that defines the
alignment requirement sort of obscures how simple the constraints
really are.  So get rid of the symbolic constants, and in place of
the ALIGN_CHECK() macro just put explicit BUILD_BUG_ON() calls with
explicit numeric values in place before each assignment.

Signed-off-by: Alex Elder <elder@linaro.org>