soc: qcom: ipa: kill ipa_context->smem_reqd_sz
The smem_reqd_sz field in the ipa_context structure is only used to
hold IPA_MEM_END_OFST, so it can be checked in ipa_pre_init() to
ensure there's enough memory available to satisfy what the software
requires.  We can just use the constant in the places the field was
used, allowing the field to go away.

Signed-off-by: Alex Elder <elder@linaro.org>