soc: qcom: ipa: assume IPA_EE_AP for event IRQ control
We only pass IPA_EE_AP as the value for the ee parameter of
gsi_event_irq_enable() and gsi_event_irq_disable().  Get rid of the
parameter and just assume IPA_EE_AP is used.

That makes it obvious that the only ee value passed to
_gsi_event_irq_control() is IPA_EE_AP, so remove that function's ee
parameter and have it assume IPA_EE_AP.

Since we're changing these function prototypes, rename them so the
"event" is at the end.

Signed-off-by: Alex Elder <elder@linaro.org>