soc: qcom: ipa: don't drop mutex in gsi_alloc_evt_ring()
We need to take the GSI mutex to protect against concurrent event
allocations.  We also need it to issue a command.

Currently gsi_alloc_evt_ring() acquires the mutex to allocate an
event, then drops it, then acquires it again before issuing the
command.  Nothing special (which might be problematic while holding
the mutex) happens between dropping it and re-acquiring it, so just
don't drop it.

Doing this means we aren't depending on the atomic behavior of the
bit operations.  In addition, mutexes have barrier semantics so we
no longer need to call smp_mb__after_atomic() when clearing the
event bit in the mask in the error path.

Signed-off-by: Alex Elder <elder@linaro.org>