soc: qcom: ipa: don't use atomic bit operations for event map
The number of bits in the event bitmap does not exceed 64, so
there's no need to treat it as a generic (arbitrary-sized) bitmap.
(In other words, we can use ffz() rather than find_first_zero_bit().)

Furthermore, we only update it while under protection of a mutex, so
there's no need to use atomic bit operations.  Just use "normal"
bitwise AND and OR operations when manipuluating bits in the event
bitmap.

Signed-off-by: Alex Elder <elder@linaro.org>