soc: qcom: ipa: don't validate interrupt mapping
In ipa_add_interrupt_handler(), there is no need to check whether
the mapping for an interrupt is valid (non-negative and less than
the max).

We can see by inspection that nothing in the ipa_irq_mapping[] array
is greater than or equal to IPA_IRQ_NUM_MAX, so there's no need to
check that.

It's possible that an interrupt isn't supported, in which case its
entyry in the ipa_irq_mapping[] array is negative.  But we know at
compile time which of these aren't supported, and none are actually
supplied to ipa_add_interrupt_handler().  So rather than checking
for that condition and returning an error, just assert that the
interrupt returned is non-negative.

Signed-off-by: Alex Elder <elder@linaro.org>