soc: qcom: ipa: modem endpoints are valid in ipa_add_interrupt_handler()
When registering a suspend interrupt, we enable the interrupt for
anything that's not a modem endpoint.  The enable mask is calculated
by clearing bits representing modem endpoints from an all-1 bitmask.

All modem clients have valid entries in the endpoint configuration
array, so checking the validity of the mapping for every modem
client's endpoint is unnecessary.  Add an assertion to make explicit
the assumption.

Signed-off-by: Alex Elder <elder@linaro.org>