soc: qcom: ipa: endpoint index is valid in ipa_get_ep_mapping()
The first thing ipa_get_ep_mapping() does is get the endpoint info
for the client it is supplied.  If the client has none, ESRCH is
returned.

A client's endpoint info is found in the ipa_ep_configuration[][]
array, and ipa_get_gsi_ep_info() returns non-null if such an entry
exists and is marked valid.  An entry exists if the client id is
non-negative and less than IPA_CLIENT_MAX.

So ipa_get_ep_mapping() can only return an endpoint mapping if the
client value is in range, and its entry in ipa_ep_configuration[][]
is marked valid.

We can see by inspection that all valid entries in that array have
ipa_ep_num values that are non-negative.  Furthermore, these values are
all less than IPA_MAX_NUM_PIPES=31 (except for IPA_CLIENT_DUMMY_CONS,
but that is considered acceptable here).

So we know that any time ipa_get_gsi_ep_info() returns non-null, its
ipa_ep_num good, so we don't need to verify the value before
returning it.

Signed-off-by: Alex Elder <elder@linaro.org>