soc: qcom: ipa: update modem consumer/producer tests
The functions IPA_CLIENT_IS_Q6_CONS() and IPA_CLIENT_IS_Q6_PROD()
are used to determine whether a given client represents a modem
consumer or producer, respectively.  They are both used in two
functions:  ipa_add_interrupt_handler() and ipa_is_modem_pipe().

Some of the values used in these functions have no valid entry
defined in the ipa_ep_configuration[][] array, and as a result,
ipa_add_interrupt_handler() reports (debug) warnings as it builds
up the mask to use for enabling the suspend interrupt.

We do not yet use any of these client types.  Removing them from
the definition of a modem consumer/producer pipe does not change
the behavior of the code either.

So stop considering pipes without valid entries in the configuration
array to be modem pipes.  Add a comment above the definition of
these two functions to clarify that the assumption about having
a valid entry in the endpoint configuration array.

Signed-off-by: Alex Elder <elder@linaro.org>