soc: qcom: ipa: client is producer in ipa_cfg_ep_mode()
The only call to ipa_cfg_ep_mode() is made conditional on the
client handle representing a producer pipe.  Change the check
in that function so it's an assertion, rather than having it
return an error code.

Rearrange the logic a bit to assign the endpoint index once based on
whether the destination is a consumer.

Signed-off-by: Alex Elder <elder@linaro.org>