soc: qcom: ipa: client handle is valid in ipa_cfg_ep()
The only caller of ipa_cfg_ep() is ipa_setup_sys_pipe().  Just prior
to the call, that function will have marked the endpoint associated
with the endpoint index supplied valid, and having done that we can
be sure the endpoint index is less than the number of pipes.

These two conditions are the only things that client_handle_valid()
checks, so we know the client handle is valid when ipa_cfg_ep() is
called.

Signed-off-by: Alex Elder <elder@linaro.org>